feat: added checkout feature for yogit
